Big Edit

+++ app/Http/Controllers/AdminController.php
@@ -10,31 +10,39 @@ use App\Models\Permissions\User;
 
 class AdminController extends Controller
 {
-    public function getRequiermemts()
+    public function getRequierments()
     {
-        $requierments = Requirement::Get();
-        if ($requierments->isEmpty()) {
-            return response()->json(['message' => "No Requirements"], 404);
+        $requierments = Requirement::all();
+        if (!$requierments) {
+            return response()->json(['message' => "Requirements Not Found"], 404);
         }
-        return response()->json($requierments, 200);
+        return response()->json(['requierments' =>$requierments], 200);
     }
 
-    public function getRequiermemt($requiermemt_id)
+    public function getRequierment($requierment_id)
     {
-        $requiermemt_info = Requirement::with('user.facility')->find($requiermemt_id);
-        if (!$requiermemt_info) {
+        if (!$requierment_id) {
+            return response()->json(['message' => "Requirement Not Found"], 404);
+        }
+
+        $requierment_info = Requirement::with('user.facility')->find($requierment_id);
+        if (!$requierment_info) {
             return response()->json(['Requirement Not Found'], 404);
         }
 
-        if ($requiermemt_info->user->facility->imgs) {
-            $requiermemt_info->user->facility->imgs = json_decode($requiermemt_info->user->facility->imgs);
+        if ($requierment_info->user->facility->imgs) {
+            $requierment_info->user->facility->imgs = json_decode($requierment_info->user->facility->imgs);
         }
 
-        return response()->json($requiermemt_info, 200);
+        return response()->json($requierment_info, 200);
     }
 
-    public function handlingRequierment(Request $request, $requiermemt_id)
+    public function handlingRequierment(Request $request, $requierment_id)
     {
+        if (!$requierment_id) {
+            return response()->json(['message' => "Requirement Not Found"], 404);
+        }
+
         $validator = validator::make($request->all(), [
             'status' => ['required', 'in:accept,reject'],
         ]);
@@ -43,18 +51,31 @@ class AdminController extends Controller
             return response()->json($validator->errors()->all(), status: 400);
         }
 
-        $requirement = Requirement::find($requiermemt_id);
+        $requirement = Requirement::find($requierment_id);
+
+        if (!$requirement) {
+            return response()->json(['message' => "Requirement Not Found"], 404);
+        }
         $user = User::find($requirement->user_id);
+
+        if (!$user) {
+            return response()->json(['message' => "User Not Found"], 404);
+        }
+
         $requirement->update([
             'status' => $request->status,
         ]);
+
         if ($request->status == 'accept') {
             $user->update(['confirmation' => '1']);
             return response()->json(['message' => "You have accepted this account"], 200);
         }
+
         if ($request->status == 'reject') {
             $user->delete();
             return response()->json(['message' => "You have deleted this account"], 200);
         }
+
+        return response()->json(['error' => "something is wrong"], 400);
     }
 }
